+++ src/Glpi/Config/LegacyConfigurators/SessionConfig.php
@@ -37,8 +37,6 @@ namespace Glpi\Config\LegacyConfigurators;
 use Glpi\Config\ConfigProviderHasRequestTrait;
 use Glpi\Config\ConfigProviderWithRequestInterface;
 use Glpi\Config\LegacyConfigProviderInterface;
-use Glpi\Debug\Profile;
-use Glpi\Debug\Profiler;
 use Glpi\Toolbox\URL;
 use Session;
 
@@ -48,15 +46,6 @@ final class SessionConfig implements LegacyConfigProviderInterface, ConfigProvid
 
     public function execute(): void
     {
-        Profiler::getInstance()->start('SessionConfig::execute', Profiler::CATEGORY_BOOT);
-        if (
-            isset($_SESSION['glpi_use_mode'])
-            && ($_SESSION['glpi_use_mode'] == Session::DEBUG_MODE)
-        ) {
-            // Start the debug profile
-            Profile::getCurrent();
-        }
-
         if (!isset($_SESSION["MESSAGE_AFTER_REDIRECT"])) {
             $_SESSION["MESSAGE_AFTER_REDIRECT"] = [];
         }
@@ -79,6 +68,5 @@ final class SessionConfig implements LegacyConfigProviderInterface, ConfigProvid
         if (isset($_REQUEST['glpilist_limit'])) {
             $_SESSION['glpilist_limit'] = $_REQUEST['glpilist_limit'];
         }
-        Profiler::getInstance()->stop('SessionConfig::execute');
     }
 }